[FIX] l10n_sa_edi_pos: display error in case of zatca or network failure
In 18.5 order validation was moved from the payment screen to `OrderPaymentValidation`.
The override of the `_finalizeValidation()` in `l10n_sa_edi_pos` was not adapted to the new
flow in the forwardport #223759 and as a result the error was not being shown anymore

This commit adapts the old override to the new order validation flow.

task-5085213

X-original-commit: 3ada1b1ccebe8873b0976d7d9ffbaf80883cbff0
